Update prev-randao tests
